Add config-flag for GDPR compliance and changed logging of ip adresses
IP Adresses now get their last octet cut of and replaced by 0 if the
new config option logging.gdpr_compliant is true. E.g 192.168.1.80
will be logged as 192.168.1.0 which provides sufficent anonymization
to comply with european data protection regulations.

If the flag is false or not present, CC will have the old behaviour
which means logging the full ips.

This flag can also be used in other places that one may find needs
adoptions to comply to EU GDPR.

Signed-off-by: Marc Misoch <47423110+MMisoch@users.noreply.github.com>